requirements.txt

tensorflow==1.9.0
keras_applications
scipy
Pillow-SIMD
h5py
